+++ packages/anthropic/src/anthropic-prepare-tools.test.ts
@@ -1,6 +1,8 @@
 import { describe, it, expect } from 'vitest';
 import { prepareTools } from './anthropic-prepare-tools';
 import { CacheControlValidator } from './get-cache-control';
+import { webFetch_20250910OutputSchema } from './tool/web-fetch-20250910';
+import { webSearch_20250305OutputSchema } from './tool/web-search_20250305';
 
 describe('prepareTools', () => {
   it('should return undefined tools and tool_choice when tools are null', async () => {
@@ -835,5 +837,91 @@ describe('prepareTools', () => {
         },
       ]
     `);
+  });
+});
+
+describe('webFetch_20250910OutputSchema', () => {
+  it('should not fail validation when title is null', async () => {
+    const problematicResponse = {
+      type: 'web_fetch_result',
+      url: 'https://test.com',
+      retrievedAt: '2025-12-08T20:46:31.114158',
+      content: {
+        type: 'document',
+        title: null,
+        source: {
+          type: 'text',
+          mediaType: 'text/plain',
+          data: '',
+        },
+      },
+    };
+
+    const schema = webFetch_20250910OutputSchema();
+
+    const result = await schema.validate!(problematicResponse);
+
+    expect(result.success).toBe(true);
+  });
+
+  it('should accept valid response with string title', async () => {
+    const validResponse = {
+      type: 'web_fetch_result',
+      url: 'https://test.com',
+      retrievedAt: '2025-12-08T20:46:31.114158',
+      content: {
+        type: 'document',
+        title: 'Example Title',
+        source: {
+          type: 'text',
+          mediaType: 'text/plain',
+          data: 'Some content',
+        },
+      },
+    };
+
+    const schema = webFetch_20250910OutputSchema();
+    const result = await schema.validate!(validResponse);
+
+    expect(result.success).toBe(true);
+  });
+});
+
+describe('webSearch_20250305OutputSchema', () => {
+  it('should not fail validation when title is null', async () => {
+    const problematicResponse = [
+      {
+        url: 'https://test.com',
+        title: null,
+        pageAge: 'April 30, 2025',
+        encryptedContent:
+          'EqgfCioIARgBIiQ3YTAwMjY1Mi1mZjM5LTQ1NGUtODgxNC1kNjNjNTk1ZWI3Y...',
+        type: 'web_search_result',
+      },
+    ];
+
+    const schema = webSearch_20250305OutputSchema();
+
+    const result = await schema.validate!(problematicResponse);
+
+    expect(result.success).toBe(true);
+  });
+
+  it('should accept valid response with string title', async () => {
+    const validResponse = [
+      {
+        url: 'https://test.com',
+        title: 'Test title',
+        pageAge: 'April 30, 2025',
+        encryptedContent:
+          'EqgfCioIARgBIiQ3YTAwMjY1Mi1mZjM5LTQ1NGUtODgxNC1kNjNjNTk1ZWI3Y...',
+        type: 'web_search_result',
+      },
+    ];
+
+    const schema = webSearch_20250305OutputSchema();
+    const result = await schema.validate!(validResponse);
+
+    expect(result.success).toBe(true);
   });
 });
